pdf-chatbot/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytest==7.4.3
python-magic==0.4.27
PyPDF2==3.0.1
celery==5.3.6
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2
